+++ gd-glm__old/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## UNRELEASED
+
+### Changed
+- Set Geostreams API url in webpack config
+  [GLM-128](https://opensource.ncsa.illinois.edu/jira/browse/GLM-128)
+
 ## [3.3.0] - 2020-06-18
 
 ### Changed